fix crash when deleting new bookmarks
(cherry picked from commit 70a2e0c31f8b0420ba4cbb94a4118e59f2d49c73)

+++ app/src/main/java/com/github/andreyasadchy/xtra/ui/channel/ChannelPagerViewModel.kt
@@ -166,7 +166,7 @@ class ChannelPagerViewModel @Inject constructor(
                             userLogin = user.channelLogin
                             userName = user.channelName
                             channelLogo = downloadedLogo }) }
-                        for (i in offlineRepository.getVideosByUserId(user.channelId.toInt())) {
+                        for (i in offlineRepository.getVideosByUserId(user.channelId)) {
                             offlineRepository.updateVideo(i.apply {
                                 channelLogin = user.channelLogin
                                 channelName = user.channelName
